+++ R/modeldb.R
@@ -9,6 +9,7 @@
 #' @importFrom dplyr all_vars group_by funs
 #' @importFrom dplyr n as_tibble filter select
 #' @importFrom dplyr case_when rename rename_all 
+#' @importFrom dplyr tbl_vars
 #' @importFrom purrr map map2 map_df transpose
 #' @importFrom purrr reduce imap pluck
 #' @importFrom tidypredict as_parsed_model